MB-11917 Async couch_view_parser implementation
Scheduler collapse in Erlang R16 is observed due to changes made
in R16 version with regard to erlang thread scheduler in general
and to scheduling NIF threads in particular. couch_view_parser is
a json stream parser NIF and takes variable amount of time to parse
incoming data depending on the incoming stream length. This resulted
in reporting improper reduction count to erlang scheduler, which in
turn caused scheduler collapse. This NIF is now implemented in
asynchronous fashion, where NIF queues the incoming tasks and returns
to erlang world immediately.

Change-Id: I9e87fe63b78b10a31aba70aba31913d60331b5d7
Reviewed-on: http://review.couchbase.org/42821
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Tested-by: Harsha H S <hhs.couchbase@gmail.com>