MB-12697 Serialize parse_json requests chunks per query context
The order of processing json parsing request chunks
was assigned randomly to an execution thread which
resulted in lexical error while processing json
parsing request during view querying. It is now
serialized, such that the request chunks per query
is processed in the same order of their arrival by
queueing every chunk to same execution thread.
However different query requests are assigned
randomly to a thread in the thread pool.

Change-Id: I16c89f7d4d422e9521ffc2a98d1d035b40e9b0fa
Reviewed-on: http://review.couchbase.org/43494
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Reviewed-by: Sriram Melkote <siri@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>