MB-12712 Use correct spatial view information
The spatial views use the native updater to update the ID b-tree and
and Erlang based updater to update the views. The native updater is
running concurrently, hence it contains outdated information about the
spatial views once it's done. Therefore are the updated views from
the Erlang based updater are used in the view group, once the native
updater is done.

Change-Id: Ic2cc53e1435c0673ddbf01d808c7638fd8de92a0
Reviewed-on: http://review.couchbase.org/43830
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Nimish Gupta <nimish@couchbase.com>