MB-12048 Make mapreduce view signature calculation clearer
The view signature calculations shouldn't ever change so that views
don't get re-created after an upgrade. In Couchbase it was just using
the #set_view{} record as a basis for the calculation. As the record
changed in 3.0, there needs to be a backwards compatible calculation
of the signature.

The existing code was a bit misleading as it seemed it would take the
b-tree and the view reference into account. But those record fields
have their initial value when the signature calculation is done. This
commit makes this fact more explicit.

Change-Id: I5ebb1d9e9d5041266b2335a7877f764d52e2dd52
Reviewed-on: http://review.couchbase.org/43833
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>