CBIDXT-255 Add compaction for spatial views
It's now possible to compact spatial views. It works in a similar way
as for the mapreduce views. If no updater is running when the compaction
is started it traverses the original data and builds a tree bottom up
from it.

If the updater starts during the compaction, then the updater will preserve
the temporary files it created (normally they get deleted after a successful
run). Those files will then be used by the compactor to apply those updates
to the already compacted file. It can happen that there are several updater
runs during one compaction, then all files of an index will be merged into
a single file.

The difference between the spatial and mapreduce views is that updates from
the temporary files into the already compacted files are done from within
Erlang and not from within Couchstore (though the temporary file creation
happens within Couchstore).

Change-Id: I15a1a6ad7f0b69c3610e57d43f1be468dd213da8
Reviewed-on: http://review.couchbase.org/45106
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Nimish Gupta <nimish@couchbase.com>