Revert "Remove the extra call for updater"
This reverts commit e799b6e7678d0c70f67e22f5cb5b030b6f2e5ee2.

The reason to revert this change is that the original commit changes
the behaviour. When there are missing partitions (e.g. during rebalance)
the stats will be updated on every retry. Though the stats should be
updated only once, as from the user perspective it was a single request
and not multiple ones.

Change-Id: Iee17a14d7f36acf4f5fe824ed5f0aaa93dabfe80
Reviewed-on: http://review.couchbase.org/47077
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Reviewed-by: Nimish Gupta <nimish@couchbase.com>
Tested-by: Volker Mische <volker.mische@gmail.com>