MB-12607 Remove the extra call to updater
We were calling the updater twice in get_set_view function to update
the stat.This was causing multiple stats call to ep-engine which was
impacting query performance in stale=false case.This change removes the
extra updater call while keeping the stat intact. After this change,
the rate almost doubled (from 29.6 to 57 with 1M items) when there were
no mutations on a single server node.

Change-Id: I7bcbc767e17ee38a8488df79e85a87d506a90d8d
Reviewed-on: http://review.couchbase.org/47190
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>