MB-13160 Update ID B-tree before de-duplication
During incremental indexing, mutations are batched and
written to temporary file and B-tree is updated whenever
a multiple snapshot is seen, or is a final batch or the
temporary file size exceeds the size limit. This is done
to optimize the disk access.

This resulted in updater writing duplicate row entries
into ID B-tree. For example, consider
..., A, B, C, D, E, #, A, E, C, #
This is a sequence of document mutations,
the `#` is a snapshot marker. Let batch size be 4.
batch-1 process: A, B, C, D which is a partial snapshot.
batch-2 process: E, #, A, E which is part of multiple snapshot.
After de-duplication of batch-2 we end up with A, (last update of) E
Force flushing at this point, as there were multiple snapshots
results in Id B-tree updated with A, B, C, D, A, E. A being duplicated.

Solution is to update B-trees from temporary files before doing
de-duplication, whenever a multiple snapshot is seen.

Add unit test for this scenario from Volker.

Change-Id: I7a36aab51193e37ab72e356280abbb704f7a8ac2
Reviewed-on: http://review.couchbase.org/48897
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Nimish Gupta <nimish@couchbase.com>
Reviewed-on: http://review.couchbase.org/51891
Tested-by: Volker Mische <volker.mische@gmail.com>