MB-14466 Mixing of data in views
We were applying a wrong rereduce function in views when we collect and merge the
query result and apply the rereduce. In that rereduce function, we create a reduce
context and store it in the erlang process table, but we were not deleting that context
processing the query. So if the next query comes and mochiweb uses the same
erlang process, the next query will find the context and use it. If both are
queries are from different views, the reduce context applied to second query
is wrong and the result will be wrong. To reproduce this issue, create following
2 view functions in different design doc on a 3 node cluster with 1000 documents:
view1:
map:
function (doc, meta) {
    emit(["A", "B"], ["couch", "base"]);
}
reduce:
function(keys, values, rereduce) {
    return ["couch", "base"];
}

view2:
map:
function (doc, meta) {
    emit(["A1", "B1"], ["couch1", "base1"]);
}
reduce:
function(keys, values, rereduce) {
    return ["couch1", "base1"];
}
Now when we query view1 and view2, we can see that view2 result
has value from view1.
Result:
{"rows":[
{"key":["A1","B1"],"value":["couch","base"]}
]
}
In this fix, now we delete the context after query is finished.

Change-Id: I37f7488136c5023e97b35fec7ca554411ebf9eb5
Reviewed-on: http://review.couchbase.org/49589
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>
Reviewed-on: http://review.couchbase.org/51854