MB-14470: Make spatial views work with active replica partitions
If some replica partitions are active, the spatial view merger failed.

The reason is, that the rows the merger is processing in that case are
already transformed into JSON, so the results are just passed on without
modification by the `make_row_fun`.

A special version of that function (compared to the mapreduce version)
is needed as the actual fold function that runs over the index is
different between spatial and mapreduce views. The spatial one has an
arity of 2, the one for mapreduce views an arity of 3 (due to the reduce).

Change-Id: I3adaa16d6b4c53f189fd582fe43dd682a5dd0082
Reviewed-on: http://review.couchbase.org/49740
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>