MB-15972 Backport MB-14601 to 3.x branch
MB-14601 update the group with correct partition seqs

If the partition state is updated in the window between a query and
before the updater has started, the updater still has stale information
of end seqs, provided by the query. In this fix, we update the new group
with correct partition seqs before returning it.

Change-Id: I87f6bdb1eb54f83f486942c8c206148ca846a469
Reviewed-on: http://review.couchbase.org/50731
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>
Reviewed-on: http://review.couchbase.org/55156
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Tested-by: Harsha H S <hhs.couchbase@gmail.com>