MB-15274 Delete old file *after* new one was renamed
The view file compaction goes through the following steps:

 1. The original file is called `*.view.1`
 2. For compaction a new file with the same name, but a `compact` extension
    is appended (`*.view.1.compact`)
 3. Once the compaction was successful, rename the new file with the number
    at the increased (`*.view.2`)

In the current code, after a successful compaction the old view file is
deleted, then the `*.compact` file is renamed. In case there would be a
crash between those two operations, there would only be a `*.compact` file
in the views directory. The view-engine won't pick it up and start re-indexing.

With this code change the old file is deleted *after* the `*.compact` file
got renamed. In case there's a crash in between those operations you will
only end up with two `*.view.<some-number>` files, but the view engine will
automatically pick up the one with the highest number.

Change-Id: I533ca5bcfd711fb967d58702e086b470421dfd48
Reviewed-on: http://review.couchbase.org/52661
Reviewed-by: Nimish Gupta <nimish@couchbase.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Tested-by: buildbot <build@couchbase.com>