MB-16385: Lower memory usage when filtering dynamically
The view-engine supports filtering out certain partitions (aka vBuckets)
on query time. This feature is used during reblance to keep the view results
stable while the underlying data is updated/cleaned up.

On large datasets it might happen that this dynamic filtering needs to touch
a lot of nodes. This leads to huge memory usage as all key-value pairs that
should be part of the final result will be kept in memory until the final
reduction happens.

The fix for this problem is to reduce the key-value pairs at an intermediate
step and only pass on the reduced value.

On my local machine the memory usage for a dataset with 10M items and filtering
out 13 out of 1024 partitions peaks at 5GB increase prior to this commit and
10MB after.

Change-Id: If873a3686978d8a4205012c666f52575954e671d
Reviewed-on: http://review.couchbase.org/55544
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Tested-by: Abhishek Singh <abhishek@couchbase.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Tested-by: Harsha H S <hhs.couchbase@gmail.com>
Tested-by: buildbot <build@couchbase.com>