[bp] MB-16826 Get rid of flawed applications starts in mochiweb.
It is flawed because it might lead to deadlocks during server
shutdown. When application:stop(App) is called, it blocks entire
application_controller process. So anything that tries to access it
from inside application being stopped will be blocked. It leads to
deadlock if such blocked call also blocks application shutdown. And
this is exactly the case with mochiweb, because corresponding
application:start calls are executed in supervisor process context
whenever mochiweb process is started.

These applications are now explicitly started by ns_server during
bootstrap.

Change-Id: I858e79812c81adf955e282ac2829069132678931
Original-Change-Id: I419c80f7bea1eb1d08094fb216891b423d502fc0
Original-Reviewed-on: http://review.couchbase.org/57096
Original-Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Original-Reviewed-by: Volker Mische <volker.mische@gmail.com>
Original-Tested-by: buildbot <build@couchbase.com>
Reviewed-on: http://review.couchbase.org/57594
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>