MB-17044: Fix race conditions when initialising a view group
When a view group is initialised a new file handler is created. That handler
also adds an entry to write guard, which makes sure that no files are opened
twice for writing. The file handler is then associated with the view group.
In case of a crash/failure the file handler is closed during the cleanup of
the view group.

There is a race condition if a crash happens *after* the file handler was
created successfully, but *before* it got associated with the group. During
the cleanup of the group, there won't be any reference to the file handler.

The solution is to clean up early and if that's not possible to have at least
a reference to the file handler in case an error occurs.

Change-Id: I0bb0ec95ffe5107243847e86ccbf941894041edc
Reviewed-on: http://review.couchbase.org/58519
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>