MB-17420: Use lowercase filenames for comparison
On Windows the different case in file paths is a problem. For example
`file:get_cwd/0` may return a path that includes an uppercase drive
letter. Though the file operation `filename:join/1` normalises the
drive letter into lowercase.

The fix is to make the full path to lowercase before comparing it.

Change-Id: I2acdd40d62d59dc3ce35ebaae7a2a245dbc5686a
Reviewed-on: http://review.couchbase.org/59012
Tested-by: Volker Mische <volker.mische@gmail.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>