MB-8512 Do not load doc when document fields are not indexed
If document is not used by all the view functions defined in design doc, we
can optimize by not loading documents during indexing.
esprima and unused java script are used to detect if javascript map
function uses document in its function body. If none of the views
defined on design doc does not index document fields in their map function,
then we prevent loading of documents for that design document during
updater runs. This helps with sync gateway views.

Change-Id: Ic3d50185cb2a5d13b29a2e43290e51b663cebf8c
Reviewed-on: http://review.couchbase.org/54777
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>