MB-18985 Use platform APIs for thread creation and synchronization
Platform APIs are compatible with NIF APIs for thread creation
and synchronization. So mapreduce NIF uses platform APIs instead
of C++11 primitivies to be compatible with NIF APIs.

On Windows, mixing C++11 primitives with NIF APIs resulted in improper
shutdown of erl.exe application in the unit tests which resulted in
error popups at the exit of every unit test run.

Change-Id: I002510632673ccbaee76f4f71b89526f00271dcd
Reviewed-on: http://review.couchbase.org/63406
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Tested-by: Harsha H S <hhs.couchbase@gmail.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>