MB-16712 Stop group synchronously after master db deletion
When the master database that contains the design documents is deleted,
the corresponding view group is also stopped/removed.

Previously that view group deletion was asynchronous. That lead to issues
when the master database was deleted, but shortly afterwards recreated. If
the view group didn't terminate fast enough (which was the case for the
spatial views in the travel sample), calls to the view group ended up in
the old group, instead of the newly created one.

We now make a synchronous call to remove the group after the master deletion,
to make sure the group got shutdown properly before any further requests
can be initiated.

Change-Id: Ib91b7000b31d9111b0cedf093de8e05c61956c1b
Reviewed-on: http://review.couchbase.org/64029
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>