MB-16737 Set default timeout late
The default timeout for internal scatter requests is 60s. It can be changed
with the `connection_timeout` query parameter. For future changes it's needed
to be able to distinguish the `connection_timeout` parameter being set at
query time or not.

Hence set the default value to `nil` and set it to the default value of
60 seconds as late as possible.

Change-Id: I45e7a2b07d27abb656150492e3a11985b77f901e
Reviewed-on: http://review.couchbase.org/64651
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>