MB-21509 [Backport] MB-21303 Handle ack from dcp producer for control_request correctly
control_request for flow control buffer size is added into the
pending request dictionary with nil as SendTo parameter as
gen_server need not reply when it gets response from producer.
This was not handled correctly when stream_response is handled
in the client which is fixed in this patch.

Change-Id: Ib382ddc9db3dc5da65d7a550d9b9ffa6ee375399
Reviewed-on: http://review.couchbase.org/68525
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Tested-by: Abhishek Singh <abhishek@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-on: http://review.couchbase.org/69313
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Tested-by: Sarath Lakshman <sarath@couchbase.com>