MB-21303 Handle ack from dcp producer for control_request correctly
control_request for flow control buffer size is added into the
pending request dictionary with nil as SendTo parameter as
gen_server need not reply when it gets response from producer.
This was not handled correctly when stream_response is handled
in the client which is fixed in this patch.

Change-Id: Iaaa2c8b62be838ed6d03fbe56c5041b2a8445607
Reviewed-on: http://review.couchbase.org/76962
Tested-by: Harsha H S <hhs.couchbase@gmail.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>