MB-21303 Handle race between stream addition and updater exit
DocLoader sub-process adds dcp stream for a vbucket.
Streams are added into active streams list after a response is
received from dcp producer. If updater exits before the stream
is added into active list, there is no way of closing this stream
until updater is stopped in subsequent iteration, which could
potentially result in not acknowledgeing any stream events for
this vbucket.

This patch avoids race by monitoring the DocLoader process and
removes the streams whenevr the process terminates or the stream
is explicitly closed.

Change-Id: I9cf738460dbba8d6adf529764e816e2c5981484c
Reviewed-on: http://review.couchbase.org/68808
Tested-by: Harsha H S <hhs.couchbase@gmail.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>