MB-21594 Remove old partition versions during rebalance
When a new vbucket is moved into a node during rebalance, for
which there is no sequence number information, we add 0 as the
start sequence number and {0, 0} as the failover log for that
partition version.
There are cases, where older failover log is still lingering in
the partition version table and this fix clears that information.

Also this results in duplicate partition as ordsets treats
{PartId, [{0,0}]} and {PartId, [{UUID1, Seq1}, {UUID2, Seq2}]}
as two different keys.

Change-Id: Idf1e669a377c365f5f1fecc57f2daabba9edec8d
Reviewed-on: http://review.couchbase.org/70389
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>