MB-24870 [Backport MB-21156] Shifting extra information to debug log level
cbcollect logs are limited to 200MB per component. The couchdb
component was logging a very high volume of information at the
info log level resulting in wrapping of the logs due to which
important information needed for triaging was lost.
This fix moves a large amount of that infomration to debug level,
which can be enabled dynamically using the REST API when the
customer faces an issue.

The API call to be used is -
curl -X POST -u username:password
	http://url:port/diag/eval
	-d 'ale:set_loglevel(couchdb,debug).

An example that worked with ./cluster_run is -
curl -X POST -u Administrator:asdasd
	http://127.0.0.1:8091/diag/eval
	-d 'ale:set_loglevel(couchdb,debug).

Change-Id: Ic7b4af748ebc2a2a2e524c43659953f1d215b8b1
Reviewed-on: http://review.couchbase.org/79607
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Tested-by: Abhishek Singh <abhishek@couchbase.com>