MB-21156 Log improvements to couchdb log
This fix makes enhancements to mitigate the loss of important
information due to rotation of the couchdb log. This includes
moving superfluous, redundant and unhelpful information to the
debug level, as well as replacing expanded lists with their
condensed form. Additionally, this improves the logging of
POST query requests by logging the ddoc and view being queried.

Change-Id: I62fea718ad24251e72746d2e4b169afbf689a8e9
Reviewed-on: http://review.couchbase.org/80987
Tested-by: Harsha H S <hhs.couchbase@gmail.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>