MB-25063 Print latest snapshot boundaries
Upon receiving a snapshot mutation, we immediately update
the Start-sequence number. However, end sequence number is
not updated, causing the "Obtaining mutations timed out"
message to print seemingly stale and misleading information.
This fix prints the latest snapshot boundaries that are
persisted when a snapshot marker is received, as well as
the Start-sequence number as reflected by the index header.

Change-Id: I03f6e0696c91746ba62fbc80c10fea26cedd36b9
Reviewed-on: http://review.couchbase.org/80698
Tested-by: Aman Kishore Achpal <aman.achpal@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>