MB-27546 Separate out ipv4 and ipv6 tests
It is observed that dcp fake server failed to start when the
bind address is changed from ipv4 to ipv6 mode.
This patch runs whole gamut of unit tests in ipv4 followed by ipv6
mode. Mode is decided by environment variable ipv6, which is
set to either true or false.

Change-Id: I5f9c6a6176f47294f78bbc56e88f807a80fd96a4
Reviewed-on: http://review.couchbase.org/92759
Reviewed-by: Aman Kishore Achpal <aman.achpal@couchbase.com>
Tested-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>