MB-27549 Fix flush
Design documents with the same View functions and Options are known
as aliases. If there are multiple aliases, they point to the same
couch_set_view_group gen_server, and share the same signature.
While flushing, we get the group_server for each {SetName, Sig} pair,
and shutdown the group_server. If there are aliases of a DDoc, then
we attempt to shutdown the same group_server multiple times, which fails
with noproc. This change catches the noproc exception.

Change-Id: I5240d03486fae1b41e2e86aad94064cffb1951b9
Reviewed-on: http://review.couchbase.org/94828
Reviewed-by: Harsha H S
Tested-by: Harsha H S