MB-32016: Disable sanitizers for Erlang dynamic modules
Some targets are incompatible with ASan / UBSan - for example library
targets which must be loaded into binaries we do not control and hence
don't link the necessary ASan / UBSan libraries. In the case of the
view engine, this includes C libraries that are loaded into beam.smp
dynamically.

For such targets utilise the new remove_sanitizers() CMake function to
disable the sanitizers; so they continue to load correctly even if the
rest of the build is built with sanitizers.

Change-Id: I4d66f9f632a739b5cf0746b64fc2494852a8f91c
Reviewed-on: http://review.couchbase.org/101898
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Sriram Melkote <siri@couchbase.com>