MB-32016: Disable sanitizers for Erlang dynamic modules (part 2)
Some targets are incompatible with ASan / UBSan - for example library
targets which must be loaded into binaries we do not control and hence
don't link the necessary ASan / UBSan libraries. In the case of the
view engine, this includes C libraries that are loaded into beam.smp
dynamically.

A previous change (38cd9b78) disabled sanitizers for mapreduce_nif and
ejson, however there are three further NIFs which need to have the
sanitizers disabled: couch_view_parser_nif, couch_ejson_compare and
snappy_nif.

Additionally the first two of those link against platform - which
itself will normally have the sanitizers enabled and hence will cause
libasan to be transitively linked (and the NIF to fail to load inside
the Erlang VM). To solve this problem, explicitly link against the
unsanitized variant of platform.

With this change 'make run-mats' successfully runs on a local Ubuntu
18.04 machine with ASan enabled.

Change-Id: Ia0a7d3e43e8818c92ce6fab1b0f6fe4bbaf8a226
Reviewed-on: http://review.couchbase.org/103280
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: <ankit.prabhu@couchbase.com>