MB-38035: Use cached decoded request body for audit message
create/update audit message needs view definition which is sent in the 
request. Since the request body is already parsed it will be cached 
in mochiweb state. If its not cached in mochiweb, it will wait for the 
response from client which will timeout since its already read 
from the client. This will hold the response till mochiweb timeout 
from reading the request body. In this patch we are cacheing the request
body when view-engine first time read the request body.

Change-Id: I663358ce50ed2da296313b23cdfd88e9228e084d
Reviewed-on: http://review.couchbase.org/122904
Tested-by: <ankit.prabhu@couchbase.com>
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Reviewed-by: Jeelan Basha Poola <jeelan.poola@couchbase.com>