MB-43900 : Address couch_audit crash if remote socket is closed.
In environments where there can flaky network, couch_audit may receive
einval while trying to estimate the remote host and port via SSL socket
object. Under the hood we make a call to ssl:peername which makes a
check whether the remote socket is still present or not.

If not, ssl:peername returns an einval which was not handled previously.
No other codepaths in couch_audit make use of ssl:peername.
Any crash in couch_audit will dump the entire request packet and the
credentials along with it.

Change-Id: I1744b95dfc7da80647762fc5869a06ced72d3294
Reviewed-on: http://review.couchbase.org/c/couchdb/+/146229
Reviewed-by: <ankit.prabhu@couchbase.com>
Tested-by: <abhishek.jindal@couchbase.com>