MB-53661: Reverse couchdb config_profile key
Reverse the couchdb configuration profile key to be (disabled, true)
in serverless mode and by default 'false' in default mode. This makes
it work naturally in config_profile:get_bool which is a wrapper around
proplists:get_bool and will naturally return false, if the key is not
found. Making it unneccessary to put the inverse key in the default
profile.

This is associated with this change in ns_server and they should be
merged to master at the same time:
https://review.couchbase.org/c/ns_server/+/180209/

Change-Id: I691e64478b22ed2e123873490f5dfb3c91b50e99
Reviewed-on: https://review.couchbase.org/c/couchdb/+/180199
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>