MB-57176: Correctly pass env to CouchDB
The config_profile data is not correctly being transferred to CouchDB's
env which causes the http server to start. Fixing that by manually
passing the data into the "env" field of the node specification. This
side of the change simply changes the call to config_profile module to
use the standard erlang API's to get the env. This is done so that we
don't share the module with CouchDB anymore and accidentally miss when
changes to that module will affect CouchDB (footgun).

Change-Id: I4a8f3ad52ba648d4934812642d146f71c13c0122
Reviewed-on: https://review.couchbase.org/c/couchdb/+/191844
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>