MB-64599: Check path is relative
mochiweb_util:safe_relative_path can return a path
that is not relative even if the original path it
got was a relative one.

filename:join() removes preceding components if any
one component is an absolute path. Ensure that when
we join AppRoot to a RelativePath, RelativePath is
indeed relative (it was absolute when
dd/../C:/path was specified).

filelib:safe_relative_path() isn't used because it
also guards against symlinks and cluster_run uses
symlinks.

Change-Id: I75036df928ed414450c5bdcbc6694217abcd1e4b
Reviewed-on: https://review.couchbase.org/c/couchdb/+/213027
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/couchdb/+/221309
Reviewed-by: <ankit.prabhu@couchbase.com>
Well-Formed: Restriction Checker