MB-64587: Merge fragmented frames
Currently there is no way of recognising that continuation frames are
expected to be sent, so frame fragmentation can't be supported. The
main options are to either include the FIN flag as part of the frame
sent to the websocket Body function, or to group the continuations
before sending to the Body function. To be consistent with other
erlang websocket implementations like cowboy, it makes sense to group
continuations into one frame.

The first frame in a fragmented message has the Fin bit 0 and opcode
non-zero (not the continuation opcode). Subsequent frames must use the
zero opcode (continuation), and the Fin bit 0 for all but the last
frame in the fragmented message. Once a frame is sent with Fin bit 1,
the fragmented message is complete, and will be sent to the client as a
single frame with the starting opcode.

A caveat to this is that the rfc states that the server must be able to
handle control messages (e.g. ping) interleaved with a fragmented
message. This requires substantial refactoring, so it is handled in a
subsequent commit.

Change-Id: I868af8f76dde4043f222b7e222519a210296dc34
Reviewed-on: https://review.couchbase.org/c/couchdb/+/222606
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>