MB-64587: Support control frames between fragment frames
Clients may send control frames (such as ping) between the frames of
a fragmented message. See section 5.4 of rfc64555.

The purpose of this functionality is to ensure that pings can be
received while a fragmented message is being received, without waiting
for the complete message to be processed. This ensures that a pong can
be sent in a timely manner.
However, the current implementation will not send the ping to the
websocket body until the surrounding fragmented message is received,
which contradicts this purpose. A later commit will ensure that
the body can handle pings without waiting for fragmented message
receiving to complete.

Change-Id: I575484fbb793b81282e51cd270c6ea38d573f2d0
Reviewed-on: https://review.couchbase.org/c/couchdb/+/222837
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>