MB-64587: Handle interleaved control frames before waiting
If control frames interleaved within a fragmented message have to wait
for the fragmented message to be completely received, they may be
handled too late, for instance in the case of ping.

Instead, when before waiting for the next packet, any complete frames
(such as ping control frames) should first be sent to the websocket
body.

Change-Id: Ie4ce1119fc9164616ffe3529d69adab283d29451
Reviewed-on: https://review.couchbase.org/c/couchdb/+/222838
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>