MB-64162: Fix header injection vulnerability
The DocId can include percent encoded characters which are not
correctly escaped in the response header.
e.g.
> PUT /couchBase/default/_design/dev_%0d%0aNewHeader:%20xxx%20%0d%0a

Response header:
< Location: http://127.0.0.1:9500/default/_design/dev_
< NewHeader: xxx

By escaping the DocId with uri_string:quote/1, we can avoid this,
instead receiving the correct Location header, and no others:

New response header:
< Location: http://127.0.0.1:9500/default/_design%2Fdev_%0D%0ANewHeader%3A%20xxx%20%0D%0A

(Note, the '/' after _design is now percent-encoded, but this doesn't
impact the meaning/functionality of the address, so this behaviour is
reasonable)

Change-Id: I5d794758d316b2f547eb945f9ffe0eabd2eeddac
Reviewed-on: https://review.couchbase.org/c/couchdb/+/223095
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>