MB-65553: Compile couchdb with rebar3
Converts couchdb compilation to use rebar3, with some necessary
changes to project structure.

A bunch of other components depend on the location of files in couchdb,
so it was decided that it wasn't worth renaming the umbrella directory
to "apps" like recommended for rebar3 projects. Instead we just use
`{project_app_dirs, ["src/*"]}` to allow rebar to correctly detect
the apps.

In order for rebar to compile each app, it needs a .app file and all
the erl files to be in a "/src" directory. For the existing
<name>.app.tpl.in files, they need to be renamed to <name>.app.src
and moved to the <name>/src directory. The remaining erl files also
had to be moved the to <name>/src directory. The .hrl files have been
moved to <name>/include for consistency, except for couch_db.hrl which
is used in too many places to be worth changing.

To avoid needing to update the include calls throughout couchdb,
couch_db.hrl is manually included in the top level rebar.config.

Since the includes for lhttpc.hrl needed updating with the new location
anyway, these were switched to the preferable include_lib call instead.

Most ERL_BUILD_OTP/ERL_BUILD macro calls have been removed from the
CMakeLists.txt files, because the REBAR macro handles compilation of
all the apps automatically.

The INSTALL macros have been updated (or added, when replacing
ERL_BUILD_OTP calls), with the new build locations in
`<build>/couchdb/_build/default/lib/`.

The layout of the files in the install directory is unchanged.

It is possible that this causes issues for running unit tests, but
these tests already do not pass, so ensuring that they still work isn't
really practical.

Change-Id: Iaaba56ccd28765182c6a1daee7afed43db88e1b5
Reviewed-on: https://review.couchbase.org/c/couchdb/+/223771
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>